requirements.txt
streamlit
torch
torchvision
Pillow
numpy
pandas
scikit-learn
matplotlib
seaborn